requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain_community>=0.0.10
langchain-groq>=0.1.0
langchain-huggingface>=0.0.10
langchain-chroma>=0.0.10
groq>=0.4.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
pypdf==3.17.1
python-magic>=0.4.27 